JavaScript re-organization
